requirements.txt
# TensorFlow for ML model inference
# Compatible with Python 3.8-3.11
tensorflow==2.12.0

# Core dependencies
numpy>=1.23.5,<2.0.0
pillow>=9.0.0

# Text-to-Speech
gtts>=2.3.0

# Google Gemini AI
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0
